Add dynamic rendering of various sensor data charts
This commit introduces dynamic rendering of charts for various sensor data on the device show page.
Instead of hardcoding charts for air temperature and wind speed,
the code now dynamically generates charts for multiple types of sensor data including:

- air temperature
- wind speed
- soil pH
- soil moisture
- rainfall
- solar radiation
- humidity

This enhancement improves the flexibility and scalability of the application
by allowing the addition of new sensor data charts with minimal code modification.

Signed-off-by: Memet Zx <zxce3@gnuweeb.org>